UPSTREAM: test_build.sh: Allow WARNERROR to be overridden
Currently, WARNERROR is hardcoded to 'yes' which is what we want most of
the time, but there are cases (such as when building with -fanalyzer)
that we don't want it enabled, so we see the entire output of the build,
and it doesn't halt at the first error.  Removing the WARNERROR line
from the script allows the variable to be overridden from an environment
variable.

Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Change-Id: Iea931e57f2a6992762566dc3dbaae8bb8df5b226
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/62745
Original-Reviewed-by: Felix Singer <felixsinger@posteo.net>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
(cherry picked from commit 9ae0463feeff61bfec90fcdb85213e3114873ad0)

Change-Id: I7ce318db6f1833121d484a66bffb9f59375510e2
Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/4313641
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Commit-Queue: Peter Marheine <pmarheine@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>