UPSTREAM: git-hooks: echo back the commit message on failure
When using nano as a git commit message editor the contents of the
commit message could get lost if the git commit hook fails due to
the lack of Signed-off-by line. To prevent this unpleasant effect
for new contributors the commit hook was modified to echo back the
commit message to allow the user to copy-paste it to the message
of the next commit attempt.

(cherry picked from commit ffc75aa710efc8796f0f34c30216ce8a2b1e3954)

Original-Change-Id: I13289f81b3fa92640aecc43eae28c1643cd7c247
Original-Signed-off-by: Miklós Márton <martonmiklosqdev@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/62879
Original-Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Original-Reviewed-by: Anastasia Klimchuk <aklm@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: ffc75aa710efc8796f0f34c30216ce8a2b1e3954
Change-Id: I9dce820f4afb50c09acc84230111731dab63011e
Signed-off-by: CopyBot <copybot.service@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/4571944
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Commit-Queue: Peter Marheine <pmarheine@chromium.org>