BACKPORT: ichspi.c: Pack hwseq_data into opaque.data field
The hwseq_data struct remains to be a stack allocated
instance for now however it is referenced in the opaque_master
data field of the flash ctx. This allows for worker functions
that leverage hwseq_data state to derive state indirectly via
the flash ctx instead of directly upon the global state.

This allows for unblocking further work to finally dispense
with the stack allocation in favour of a heap allocation with
a defined life-time of the driver.

BUG=b:237839418
TEST=builds

Original-Change-Id: I2ad57d496176cf26edcd0ba40154cc6250846e33
Original-Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/65205
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Anastasia Klimchuk <aklm@chromium.org>
Original-Reviewed-by: Alexander Goncharov <chat@joursoir.net>
Original-Reviewed-by: Thomas Heijligen <src@posteo.de>
(cherry picked from commit 80f1ea9a791cef897f1597d11e0c37703a868cd7)

Change-Id: Iaa817c54fb48d23a83ccbdb4f79c350c39f14289
Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3958299
Reviewed-by: Evan Benn <evanbenn@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>