BACKPORT: meson: Build the ich_descriptors_tool as feature
Provide the meson option `-Dich_descriptors_tool=auto/enable/disabled`
to determin if the ich_descriptors_tool should be build or not. On
`auto` or `enabled` it will be, on `disabled` it will not be build.
This is usefull for environments where the ich_descriptors_tool is not
needed.

Backport needed because meson.build file in cros tree had commented
line `#subdir('util')` which is now replaced by the contents of this
patch. The line was commented a while ago because at that time util/
directory could not be built with meson, but now it does.

BUG=none
BRANCH=none
TEST=emerge

(cherry picked from commit e4137c6913c7a84e5a4a8e0b4fd571d5d0b78616)

Cq-Depend: chromium:3925170
Original-Change-Id: Ief65a914019f72536d563eda36bd7f2f5330bea8
Original-Signed-off-by: Thomas Heijligen <thomas.heijligen@secunet.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/66704
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Change-Id: I13443852ba8b8a44eee7b8d9a75137c34938ea4d
Signed-off-by: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3920380
Reviewed-by: Peter Marheine <pmarheine@chromium.org>