BACKPORT: tree/: Convert flashchip probe func ptr to enumerate
This forges the way for flashchips.c to be pure declarative
data and lookup functions for dispatch to be pure. This
means that the flashchips data could be extracted out to
be agnostic data of the flashrom code and algorithms.

Backport is needed to convert Rex chip entry in flashchips.

BUG=b:242466990
BRANCH=none
TEST='R|W|E && --flash-name' on ARM, AMD & Intel DUT's.

(cherry picked from commit 10e7a0ebd74db72f00b8fc1b741f411d5a49b92d)

Original-Change-Id: I00aaab9c83f305cd47e78c36d9c2867f2b73c396
Original-Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/66781
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nikolai Artemiev <nartemiev@google.com>
Original-Reviewed-by: Felix Singer <felixsinger@posteo.net>
GitOrigin-RevId: 10e7a0ebd74db72f00b8fc1b741f411d5a49b92d
Change-Id: Ic66d8e1f20b5974c28a18c460394dcc4922fb643
Signed-off-by: CopyBot <copybot.service@gmail.com>
Signed-off-by: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3996288
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>