BACKPORT: tree/: Convert flashchip read func ptr to enumerate
This forges the way for flashchips.c to be pure declarative
data and lookup functions for dispatch to be pure. This
means that the flashchips data could be extracted out to
be agnostic data of the flashrom code and algorithms.

Backport is needed to convert the Rex chip entry in flashchips.
Second reason is diffs in flashrom.c in the functions handling
read and verify operations.

BUG=b:258095849
BRANCH=none
TEST='R|W|E && --flash-name' on ARM, AMD & Intel DUT's.

Original-Change-Id: I612d46fefedf2b69e7e2064aa857fa0756efb4e7
Original-Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/66788
Original-Reviewed-by: Nikolai Artemiev <nartemiev@google.com>
Original-Reviewed-by: Felix Singer <felixsinger@posteo.net>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
(cherry picked from commit 594d3357b6a4d883b6b87295bf4b9b6400777d18)
Change-Id: I46adb4f6a74dda8865476ab2b6ae211495c59536
Signed-off-by: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/4011603
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>