UPSTREAM: programmer_init: Work on a mutable copy of programmer params
The signature of extract_param() was wrong all the time. It actually
modifies the passed, global `programmer_param` string. This only com-
piled w/o warnings because of a deficiency of the strstr() API. It
takes a const string as argument but returns a mutable pointer to
a substring of it.

As we take a const parameter string in the libflashrom API and should
not change that, we create a copy in programmer_init() instead.

Now that we free our copy of the programmer parameters at the end of
programmer_init() it's more obvious that the string can only be used
during initialization. So also clear `programmer_param` inside
programmer_init() instead of programmer_shutdown().

(cherry picked from commit ca61c3dd3efcdf1aa5c7ea055514b0985e5fff48)

Original-Change-Id: If6bb2e5e4312b07f756615984bd3757e92b86b0a
Original-Signed-off-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/67094
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Original-Reviewed-by: Thomas Heijligen <src@posteo.de>
GitOrigin-RevId: ca61c3dd3efcdf1aa5c7ea055514b0985e5fff48
Change-Id: Ibaa206bf5b42e248ad536bd4bc072cfbe06188f7
Signed-off-by: CopyBot <copybot.service@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3903339
Commit-Queue: Peter Marheine <pmarheine@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Peter Marheine <pmarheine@chromium.org>