CHROMIUM: flashchips.c: Fake flash part size as 32MB for Rex Proto 0
Rex Proto 0 has 64MB flash part but the size of image file is 32MB.
Flashrom tool fails to program an image on size mismatch.
Increasing the size of image file will increase the time to program
the flash part which involves the steps like read/write/verify so
this patch tries fake the flash part size as 32MB.

This is a temporary change for Rex Proto 0 and shall be reverted for
Proto 1 which will have 32MB flash part.

BUG=b:242958619,b:243615601
TEST=Verified by flashing 32MB image on Google Rex.

Signed-off-by: Kapil Porwal <kapilporwal@google.com>
Change-Id: I3441f8e211b5ceffdefb5cc0a1f8134fe26c0425
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/3897831
Tested-by: Kapil Porwal <kapilporwal@chromium.org>
Commit-Queue: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Tarun Tuli <taruntuli@google.com>
Tested-by: Subrata Banik <subratabanik@chromium.org>