UPSTREAM: Makefile: Don't install git hooks automatically
These specific git hooks are only needed when someone wants to push a
patch to upstream and so it's not needed to run it in every make call.
Beside that, we also don't know the environment in which this is
executed and it might not be wanted.

Thus, add a new make target `gitconfig` and move the install command to
it. It can be used by running `make gitconfig`.

(cherry picked from commit 39b189077379a6cdd99e5ae20452fa685b94500e)

BUG=none
BRANCH=none
TEST=builds

Original-Signed-off-by: Felix Singer <felixsinger@posteo.net>
Original-Change-Id: Ib83568c7ff149a8ec34ad7e92720c36a89def7bd
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/68647
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Original-Reviewed-by: Anastasia Klimchuk <aklm@chromium.org>
GitOrigin-RevId: 39b189077379a6cdd99e5ae20452fa685b94500e
Change-Id: I268005b593dbaa0feda21c4a5d677978d2b718d2
Signed-off-by: CopyBot <copybot.service@gmail.com>
Signed-off-by: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/4004982
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>