flashrom.c: Import upstream erase_by_layout() impl
In order to switch the implementation of erase to the upstream
implementation and drop the cros variant we need the ability to
A/B test both implementations in-tree. Import the upstream
implementation and allow for easy switching to shakedown
the upstream variant.

BUG=b:260313170
BRANCH=none
TEST=build

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: I4da039a11e2638676a366acf82c572248e4c9eb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/4142018
Reviewed-by: Sam McNally <sammc@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>