flashrom.c: Import upstream write_by_layout() impl
In order to switch the implementation of write to the upstream
implementation and drop the cros variant we need the ability to
A/B test both implementations in-tree. Import the upstream
implementation and allow for easy switching to shakedown
the upstream variant.

BUG=b:260313170
BRANCH=none
TEST=build

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: I27f6627bb0287fe41174329b71d0ad22f7cad62e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/4142019
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>