UPSTREAM: tests/chip: Add non-aligned write within a region unit-test
A written region that is sized below that of the erasure granularity
can result in a incorrectly read region that does not include prior
content within the region before the write op. This was dealt with
in ChromeOS downstream by expanding out the read to match the erase
granularity however does not seem to impact upstream. Add a unit-test
to avoid regression as this is important behaviour to cover.

(cherry picked from commit 3d76865fd50ab9b28f367ed27b19d8746548cdb1)

Original-Change-Id: Id3ce5cd1936f0f348d34a6c77cee15e27a5c353f
Original-Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/71659
Original-Reviewed-by: Sam McNally <sammc@google.com>
Original-Reviewed-by: Evan Benn <evanbenn@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 3d76865fd50ab9b28f367ed27b19d8746548cdb1
Change-Id: I24ee6a35aea1cdd48597497762bcc85b54aea48d
Signed-off-by: CopyBot <copybot.service@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/4172451
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>