BACKPORT: tests: Don't require old unlock function to be called
This is a partial cherry-pick of `commit
00c9fac2a66ab2aafc6949166dfc733fd0b747ea` that only includes unit test
changes. The unlocking changes in the original commit have been removed
because they caused a performance regression and have been reimplemented
in a later patch.

This commit deletes part of a test that required the chip unlock
function to be called before read/write/erase operations because WP
unlocking is now used instead of chip unlocking.

BUG=b:280111380,b:283779258
BRANCH=none
TEST=ninja test

Change-Id: I7a733fa09c163f81970ea948226b5defda40a540
Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/4680094
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>