UPSTREAM: doc: Fix the example Signed-off-by line to be displayed properly
It was not displayed, so needed to be fixed. Secondly, it is just
one line and can be an inline code markup instead of a code-block
(the latter was not displayed for some reasons).

(cherry picked from commit cc145f0a21ffe11f06f6ddc819ad274ade782e74)

Original-Change-Id: I0f26d6fb03d1491daefa4a362068f674230e94a6
Original-Signed-off-by: Anastasia Klimchuk <aklm@flashrom.org>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/76663
Original-Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: cc145f0a21ffe11f06f6ddc819ad274ade782e74
Change-Id: I0f37d8c850dad151674e56f20ec9b89d5d7e468f
Signed-off-by: chromeos-ci-prod <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/4724746
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Commit-Queue: Peter Marheine <pmarheine@chromium.org>