UPSTREAM: ichspi: Change the opcode position for reprogramming on the fly 2->4
The function for reprogramming on the fly was using the default
configuration O_ST_M25P as a base and the position 2 as the position
to reprogram. Position 2 corresponds to JEDEC_SE which is often useful
for chip erase (when less than whole chip needs to be erased).

This patch changes the default position to reprogram to 4, which
corresponds to JEDEC_REMS. It is used less often, but if it needs to
be used, it will be discovered missing and reprogrammed back.

For erase opcodes, there is usually several of them available. So if
one is missing, erase still can be performed with the remaining ones.
However, this hides the fact that one of available erase opcodes is
missing (it won't be reprogrammed back), and also it gives non-optimal
erase layout.

(cherry picked from commit f5d1bed6d4132372fe303b2d3a1d11295626b2ea)

Original-Context: https://ticket.coreboot.org/issues/556
Original-Change-Id: I6bc855daedf0af2e8de191f23a3512de3ebc3fef
Original-Signed-off-by: Anastasia Klimchuk <aklm@flashrom.org>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/84567
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nikolai Artemiev <nartemiev@google.com>
GitOrigin-RevId: f5d1bed6d4132372fe303b2d3a1d11295626b2ea
Cr-Build-Id: 8732568628815896817
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8732568628815896817
Copybot-Job-Name: flashrom-main-copybot-downstream
Change-Id: I5a5eb5b739177aafdefe53fba8a94c6949175f97
Signed-off-by: chromeos-ci-prod <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/5920173
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>
Tested-by: Hsuan Ting Chen <roccochen@chromium.org>