UPSTREAM: ichspi: Check whether chipset is locked when probing opcode
This is follow up (or fix) for
commit 26a1eb514ccefc61b110068cf0eea73c397ba045

When probing opcode, all opcodes in POSSIBLE_OPCODES are reported as
supported, even if the opcode is not in curopcodes. This is relying
on reprogramming on-the-fly to handle the gap between POSSIBLE_OPCODES
and curopcodes. However, for locked chipsets on-the-fly is not
happening (is not possible, since list of opcodes is locked), so we
can't rely on it.
So, we need to check whether chipset is locked.

(cherry picked from commit b184e013e2b7c8e4363d74c72fe9c8cc222d519b)

Original-Ticket: https://ticket.coreboot.org/issues/573
Original-Change-Id: I701a86f030cfef43a1158bf075287ade569254e6
Original-Signed-off-by: Anastasia Klimchuk <aklm@flashrom.org>
Original-Tested-by: Matt DeVillier <matt.devillier@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/85592
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nikolai Artemiev <nartemiev@google.com>
Original-Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Original-Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>
GitOrigin-RevId: b184e013e2b7c8e4363d74c72fe9c8cc222d519b
Cr-Build-Id: 8718975309589020833
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8718975309589020833
Copybot-Job-Name: flashrom-main-copybot-downstream
Change-Id: I3083ca5a0598e4e1adba454f8ccef25570ddb61e
Signed-off-by: chromeos-ci-prod <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/6104722
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>