CHROMIUM: flashchips.c: Fake flash part size as 32MB for Felino proto 0
Felino Proto 0 has 64MB flash part but the size of image file is 32MB.
Flashrom tool fails to program an image on size mismatch.
Increasing the size of image file will increase the time to program
the flash part which involves the steps like read/write/verify so
this patch tries fake the flash part size as 32MB.

This is a temporary change for Felino Proto 0 and shall be reverted for
Proto 1/EVT which will have 32MB flash part.

BUG=b:409472563,b:412502949
TEST=Verified similar changes on previous generation board.

Change-Id: I5947031a0230e31a20b63ebcad9c5753fadfd5e4
Signed-off-by: Kapil Porwal <kapilporwal@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/6474988
Tested-by: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>
Commit-Queue: Kapil Porwal <kapilporwal@chromium.org>