CHROMIUM: util/flashrom_tester: Make 'built' dependency optional
The 'built' crate provides detailed build-time information but add a
dependency and compilation overhead. This change makes it optional to
allow for more minimal builds where this information is not required.

A new 'build-info' feature is introduced to control this functionality.
It is enabled by default to maintain the existing behavior for standard
builds.

BUG=b:426471570, b:389852236
TEST=sudo emerge -1 flashrom-tester
TEST=(HOST)flashrom_tester --version

Change-Id: I7a400518cb87c44ff08f34b5cd4174520c91ecb4
Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/6656624
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>
Commit-Queue: Jian-Jia Su <jjsu@chromium.org>