CHROMIUM: util/flashrom-tester: Allow the output stream to be injected in logger
The current Logger is tightly coupled to std::io::stdout, which makes
unit testing cumbersome. It requires using the 'gag' crate to capture
and inspect standard output.

This CL refactors the Logger to be generic over a W: Write + Send type,
allowing the output stream to be injected as a dependency.
* The Logger struct now holds a writer: Mutex<W>.
* The init() function provides std::io::stdout for production use.
* Unit tests now inject a Vec<u8> buffer, removing the need for the gag
  crate and simplifying the test setup.

BUG=b:426471570
TEST=`cargo test` in util/flashrom_tester

Change-Id: I1cc647ad1500ae70ef5061b1ef0505b332fd1d3a
Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/6817923
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>