CHROMIUM: bindings/rust/libflashrom: Remove gag dependency from tests
This CL removes the gag crate from the libflashrom tests.

The logging_stderr test has been updated to use the existing test_logger
infrastructure, which writes log output to a thread-local buffer. This
approach is more direct and idiomatic for testing logging functionality,
as it avoids the need to capture stderr.

Additionally, the original logging_stderr test was repurposed. Its
implicit goal was to verify the default log_eprint() function. Given the
simplicity of this function, a dedicated integration-style test
capturing stderr is unnecessary.

This test has been renamed to logging_log_level to better reflect its
new purpose: to verify the log level filtering.

BUG=b:426471570
TEST=`cargo test` in bindings/rust/libflashrom

Change-Id: I44635d089060ec7093439b52a8efb5509e4bb0c3
Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/6817925
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>