CHROMIUM: util/flashrom_tester: Make 'sys-info' dependency optional
The 'sys-info' crate is used for a single purpose: to check the
operating system release in the 'host_is_chrome_test'. This test is
specific to ChromeOS environments and is not relevant on other
platforms. Making the dependency optional improves the tester's
portability and allows it to be built and run on non-ChromeOS systems
without requiring this platform-specific check.

Create another feature flag, chromeos-host, for this.

BUG=b:426471570
TEST=`cargo test` in util/flashrom_tester
TEST=`cargo test --no-default-features`

Change-Id: I2e63725fc8a8ab36a7ef1800f7512a10056ffdff
Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/flashrom/+/6817762
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>